Moved plural and singular constants to global
