+++ app/javascript/src/components/commons/MenuBar/index.jsx
@@ -4,7 +4,7 @@ import { Typography } from "neetoui";
 import { MenuBar as NeetoUIMenuBar } from "neetoui/layouts";
 import { useTranslation } from "react-i18next";
 
-import { PLURAL } from "components/Dashboard/constants";
+import { PLURAL } from "constants";
 
 import { SEGMENTS_ICON_PROPS, TAGS_ICON_PROPS } from "./constants";
